backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-community
langchain-huggingface
langchain-openai
faiss-cpu
sentence-transformers
python-dotenv
sqlalchemy
python-multipart
pypdf
lark-oapi